+++ lib/src/distribution/maps/uniform_distribution_map.dart
@@ -1,6 +1,6 @@
 import 'package:inspire_blur/src/distribution/blur_distribution_map.dart';
 
-class UniformDistributionMap extends BlurDistributionMap {
+class UniformDistributionMap extends IntensityBasedDistributionMap {
   const UniformDistributionMap({
     required super.width,
     required super.height,