+++ test/config_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:inspire_blur/src/distribution/blur_distribution.dart';
 import 'package:inspire_blur/src/inspire_blur_config.dart';
+import 'package:inspire_blur/src/transform/blur_transform.dart';
 
 import 'test_core.dart';
 
@@ -14,6 +15,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       );
 
   test('Constructor accepts valid configs', () {
@@ -25,6 +27,7 @@ void main() {
         stops: [0, 1],
         values: [0, 1],
       ),
+      transform: BlurTransform.identity,
     );
 
     final configWithSigmaX = InspireBlurConfig(
@@ -35,6 +38,7 @@ void main() {
         stops: [0, 1],
         values: [0, 1],
       ),
+      transform: BlurTransform.identity,
     );
 
     final configWithSigmaY = InspireBlurConfig(
@@ -45,6 +49,7 @@ void main() {
         stops: [0, 1],
         values: [0, 1],
       ),
+      transform: BlurTransform.identity,
     );
 
     final configWithSigmaXY = InspireBlurConfig(
@@ -56,6 +61,7 @@ void main() {
         stops: [0, 1],
         values: [0, 1],
       ),
+      transform: BlurTransform.identity,
     );
 
     expect(configWithSigma.sigma, equals(1));
@@ -88,6 +94,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ),
       throwsAssertionError,
     );
@@ -105,6 +112,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ),
       throwsAssertionError,
     );
@@ -120,6 +128,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ).copyWith(
         sigmaX: 5,
         sigmaY: 5,
@@ -139,6 +148,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ),
       throwsAssertionError,
     );
@@ -154,6 +164,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ).copyWith(sigmaX: 5),
       throwsAssertionError,
     );
@@ -170,6 +181,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ),
       throwsAssertionError,
     );
@@ -185,6 +197,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ).copyWith(sigmaY: 5),
       throwsAssertionError,
     );
@@ -279,6 +292,7 @@ void main() {
         begin: Alignment.topCenter,
         end: Alignment.bottomCenter,
       ),
+      transform: BlurTransform.identity,
     );
     final copiedConfig = config.copyWith();
 
@@ -296,6 +310,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ).copyWith(sigmaX: 5),
       throwsAssertionError,
     );
@@ -308,6 +323,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ).copyWith(sigmaY: 5),
       throwsAssertionError,
     );
@@ -320,6 +336,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ).copyWith(sigma: 5),
       throwsAssertionError,
     );
@@ -332,6 +349,7 @@ void main() {
           stops: [0, 1],
           values: [0, 1],
         ),
+        transform: BlurTransform.identity,
       ).copyWith(sigma: 5),
       throwsAssertionError,
     );